VMM/GIC: bugref:10877 Fixes for pedantic sign-compare warnings on GCC [part 2]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108732 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMAll/GITSAll.cpp
@@ -59,7 +59,7 @@
 #ifdef LOG_ENABLED
 DECL_HIDDEN_CALLBACK(const char *) gitsGetCtrlRegDescription(uint16_t offReg)
 {
-    if (offReg - GITS_CTRL_REG_BASER_OFF_FIRST < GITS_CTRL_REG_BASER_RANGE_SIZE)
+    if ((uint32_t)offReg - GITS_CTRL_REG_BASER_OFF_FIRST < GITS_CTRL_REG_BASER_RANGE_SIZE)
         return "GITS_BASER<n>";
     switch (offReg)
     {
